requirements/card-age-report.txt
asn1crypto==0.24.0
backoff==1.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
ghzh-clients==0.2.0.dev0
github3.py==1.3.0
idna==2.8
Jinja2==2.10.1
jwcrypto==0.6.0
MarkupSafe==1.1.1
pycparser==2.19
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
